refactor: Separate Navigation component from App component
