+++ src/App.scss
@@ -1,7 +1,3 @@
 .App {
   text-align: center;
 }
-
-ol {
-  display: inline-block;
-}
