+++ src/components/navigation/Navigation.scss
@@ -1,0 +1,3 @@
+ol {
+    display: inline-block;
+}
