+++ src/models/contents.tsx
@@ -1,0 +1,2 @@
+export type Topic = { id: number; title: string; body: string; };
+export type AppMode = 'READ' | 'WELCOME' | 'CREATE';
